Extract MimeGenerator
